+++ packages/core-js/modules/esnext.symbol.replace-all.js
@@ -1,5 +1,4 @@
+// TODO: remove from `core-js@4`
 var defineWellKnownSymbol = require('../internals/define-well-known-symbol');
 
-// `Symbol.replaceAll` well-known symbol
-// https://tc39.github.io/proposal-string-replaceall/
 defineWellKnownSymbol('replaceAll');